.history/requirements_20240912143949.txt
# generated by rye
# use `rye lock` or `rye sync` to update this lockfile
#
# last locked with the following flags:
#   pre: false
#   features: []
#   all-features: false
#   with-sources: false
#   generate-hashes: false
#   universal: false

-e file:.
altair==5.4.1
    # via streamlit
attrs==24.2.0
    # via jsonschema
    # via referencing
blinker==1.8.2
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via click
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.8
    # via requests
jinja2==3.1.4
    # via altair
    # via pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.6.4
    # via altair
numpy==2.1.1
    # via pandas
    # via pyarrow
    # via pydeck
    # via streamlit
packaging==24.1
    # via altair
    # via streamlit
pandas==2.2.2
    # via streamlit
pillow==10.4.0
    # via streamlit
protobuf==5.28.0
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via smatsu-st-2
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via jsonschema
    # via jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.8.1
    # via streamlit
rpds-py==0.20.0
    # via jsonschema
    # via referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.38.0
    # via smatsu-st-2
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
typing-extensions==4.12.2
    # via altair
    # via streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.2
    # via streamlit
